+++ Assets/Scripts/Control/SceneObjectControl/BulletControl.cs
@@ -37,8 +37,8 @@ public class BulletControl : MonoBehaviour
             octave = 8,
             intensity = 10,
             sizeRange = Vector2.up * 0.1f,
-            colorIndex = (int)ColorIndexMode.HighlightToAlpha,
-            sizeIndex = (int)SizeCurveMode.Small_Hight_Small,
+            colorIndex = ColorIndexMode.HighlightToAlpha,
+            sizeIndex = SizeCurveMode.Small_Hight_Small,
             textureIndex = 0,
             groupCount = 1,
         };
