+++ Assets/Scripts/Effect/Pooling/Bullet/Bullet_Pooling.cs
@@ -42,8 +42,8 @@ public class Bullet_Pooling : ObjectPoolBase
             octave = 8,
             intensity = 20,
             sizeRange = sizeRange,
-            colorIndex = (int)ColorIndexMode.HighlightToAlpha,
-            sizeIndex = (int)SizeCurveMode.Small_Hight_Small,
+            colorIndex = ColorIndexMode.HighlightToAlpha,
+            sizeIndex = SizeCurveMode.Small_Hight_Small,
             textureIndex = 0,
             groupCount = 1,
         };
@@ -69,8 +69,8 @@ public class Bullet_Pooling : ObjectPoolBase
             octave = 8,
             intensity = 20,
             sizeRange = sizeRange,
-            colorIndex = (int)ColorIndexMode.HighlightToAlpha,
-            sizeIndex = (int)SizeCurveMode.Small_Hight_Small,
+            colorIndex = ColorIndexMode.HighlightToAlpha,
+            sizeIndex = SizeCurveMode.Small_Hight_Small,
             textureIndex = 0,
             groupCount = 1,
         };
